requirements.txt
asgiref==3.8.1
boto3==1.35.94
botocore==1.35.94
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-ses==4.3.1
djangorestframework==3.15.2
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
